suppression of redundant code
