+++ packages/yourloops/test/integration/pages/patient-view/trends/trends-view-mobile.patient.spec.tsx
@@ -34,30 +34,13 @@ import { patient2Info } from '../../../data/patient.api.data'
 import { mockWindowResizer } from '../../../mock/window-resizer.mock'
 import { AppUserRoute } from '../../../../../models/enums/routes.enum'
 import { getMinimalTrendViewData } from '../../../mock/minimal-trend-view-data'
-import mediaQuery from 'css-mediaquery';
-
-function mockScreenWidth(width: number): void {
-  globalThis.matchMedia = (query: string): MediaQueryList => ({
-    matches: mediaQuery.match(query, { width }),
-    media: query,
-    onchange: null,
-    addListener: () => {
-    },
-    removeListener: () => {
-    },
-    addEventListener: () => {
-    },
-    removeEventListener: () => {
-    },
-    dispatchEvent: () => true
-  });
-}
+import { mockMobileScreen } from '../../../mock/mobile-screen.mock'
 
 describe('Trends view for patient', () => {
   beforeEach(() => {
     mockWindowResizer()
     mockPatientLogin(patient2Info)
-    mockScreenWidth(400)
+    mockMobileScreen()
   })
 
   afterEach(() => {
